Add __init__.py files, turning this into a module
I've decided that the things I build for Maya are just not going to be
proper Python packages. It's just too much work, especially when one
relies on another, and it doesn't match up at all with how we use code
as tech artists in Maya. I want to have submodules that can be
relatively imported, at any hash, and I want to update all of it all day
long, while working on libraries in Maya. Python packaging is a total
nightmare, and I'm tired of being beholden to bad design. My choices
here don't indicate a fix to the bad design, and bring their own issues,
but I know these issues better, and the effort required on my end for
passable results is a fraction of the Python way.