Unify path dict levels to (name, part, children)
I had to give up on my idea that the root path part[s] had no name[s],
to bring all levels of the passed-in dict into uniformity. Now it's all
identical triples of the name for the part, the part added at that
level, and the list of children, which can be empty. This will make
walking the tree trivial. This also makes fst and snd less useful, as
we're no longer working with pairs, so their gone.