Add path table building
This was called flatten for a moment there, and that's the idea. This
new .buildPathTable method walks the tree, and creates a single-level
dict from the names tagging each level, each pointing to the fully
resolved path. This allows the main use of this module's currently only
class, which is to tag interesting parts of the project's file system
for easy retrieval, all throughout tools in Maya.