Revamp util.parseHier's normalization routine
I realized that the level parsing work I was doing in mpath.parseLayout
was something I could do more easily at the string hierarchy parsing
level (from which, parseLayout will later benefit). This manifested as a
rewrite of parseHier's normalization routine, and meant culling most of
mpath.parseLayout, and the moving over of its tests.

There was a problem with how I was determining levels, illustrated here:

a
    b
        c
    d
      e
      f
    g

In the old routine, b, d, and g would be considered level 2, e and f,
level 3, and c, level 4, where, clearly, c, e, and f should all be
considered level 3, under their respective parents. I realized each
element [probably shouldn't, but] could have its own, internal
indentation levels, and one way to see this follows here:

Let's say we have this valid hierarchy:

b
    c
    d
        e

...and this, also valid hierarchy:

f
   g
    h
    i
   j
    k

It makes sense to be able to merge 2 such valid trees into a new tree:

a
    b
        c
        d
            e
    f
       g
        h
        i
       j
        k

The b and f nodes were, and still are their own little trees, and, can
have their own internal indent rules, even after they've been made
siblings. There should be no confusion between c and d, and h, i, and k,
which are indented the same, but are clearly not on the same level.

The only real requirement of indentation is that everything respects the
outdent rule, i.e. that you can't outdent to a level that doesn't exist
above you in the level stack, example here:

a
    b
  c

Level c is floating between worlds. We could "flatten" or "round" levels
down, and consider c to be on the level of a, but, unlike new indent
levels under a given level, detailed above, this one feels somehow more
ambiguous. It's something the Python parser disallows, with an error,
and I've decided I agree that outdents shouldn't be allowed to do this.