Add mpath.parseLayout, test
This was a mess of fighting with imperative code over a 4-day Memorial
Day weekend. My heart wasn't in it, and I did it all wrong. I should
have played with the ideas in types/Haskell. I should have TDD'd it for
each little idea, so I'd have tests in place for level increases,
decreases, etc., so I was more free to play with ideas... Instead, I
wrote one, unfinished test, and then wrestled with the entire algo for
days, wasting a bunch of time for who knows what reason. This should
have been simple. Alas.