Add path existing checking to mpath.Path instances
I was using tempfile.tempdir[1], but experiencing a very strange problem
in the tests, if the test_name began with 'a' through 'k'. Anything from
'l' through 'z' would work fine. So, for instance, if I named the test
"test_k", it would fail, but "test_m" would succeed (this took forever
to track down), and this was due to tempfile.tempdir being set to None
in the former, and "/tmp" in the latter. I found gettempdir[2] in the
helps, and that fixed the issue.

[1] https://docs.python.org/2.7/library/tempfile.html#tempfile.tempdir
[2] https://docs.python.org/2.7/library/tempfile.html#tempfile.gettempdir