Define __str__ for mpath.MPaths
In my first use of this lib in Maya, I immediately saw that evaluating
an MPaths in the Maya Script Editor showed a bunch of Path objects in
the printed out dict, and str didn't help; it only stringified the Path
values in the output. So, I rolled my own string, which sorts by key.