Derive Path from str
I didn't want to do this. Now, evaluating an MPaths in the Maya Script
Editor prints the strings of the paths in the dict, which looks nicer,
but also hides what they really are a bit deeper, which I'm not entirely
in love with. At least repr() still prints them out, so I guess that's
good. The real reason I did this, though, was because, though it claims:

- # Error: TypeError: file <maya console> line 1: coercing to Unicode: need string or buffer, Path found #

...the `with open()` context would not coerce a simple mpaths["path"] to
unicode, even if I implemented __unicode__, and tried every variation of
returning unicode(self) and unicode(str(self)), and trying to set the
__repr__ method equal to __str__ and __unicode__... The final option was
to just let Path be a str, just as MPaths is a dict. All tests pass!