Add trailing slashes to dirs in MPaths.pformat
This is just a nice-to-have, and it only works if the files exist on
disk at the time it's called. It makes the check for existence, then to
see if it's a directory, and if so, tacks on a "/", else does nothing.
This also saw me rewriting the code from functional/lambda style to more
of a full-blown function, albeit nested for currying via map.