Add MPaths.mayaForm
This will allow overriding of paths passed in by the otherwise
immutable, team-shared defaults, passed in to a given tool by, say, a
menu creator script. This way, you're not locked down to the defaults,
if you need to go off-road for a bit, and tools-writers also won't leave
you having to dig through scripts and configs, looking for where to set
some overrides, and you won't have to, say, edit a menu script locally,
without checking it in, just so a tool you launch will do what you want.

This is just a display at the moment; you can edit the fields, but
nothing happens. Nothing gets stored.

Here's where I want to take this:

    - changes to the fields are written to optionVars
    - fields turn red if they contradict the defaults
    - some kind of UI indication of paths that don't exist
    - roll over fields for pop-up info
        - fully qualified paths
        - original part, if modified
    - a way/ways to print/pretty-print out the information
    - righ-click in a modified field for reset to default option
        - prints out "changed MPaths part 'name' from 'old' to 'new'
            - like how git prints out changes, so you can undo them
    - delete/add paths(?)

For simplicity, I think any changes would just be concated back up as a
string, passed back into the system to re-parse, with the UI then being
fully refreshed.

I also want some kind of nominal system for identifying the stored
information as belonging to a given UI, probably by passing in a name
for the set of paths when creating an MPaths instance, which would be
used to create namespace prefixed opVars (a technique I've used in the
past). This would also allow finding all related opVars for various
cleanup needs.