Fix bug in util.parseStrHier implementation
I didn't read the errors closely enough, and spent too much time
debugging mpath.parseLayoutStr, when the bug was really in this
util.parseStrHier function. That's why the test is duplicated across
them, but another test can't hurt, so I'm leaving it in at both levels.

The bug showed up not when I indented, outdented, then indented, but
when I did those 3, and then outdented again! What a rascal.