Stop MPaths deriving from dict
I didn't like how deriving from dict ballooned the output from help(),
and I couldn't see a real reason for it to have to be a dict, and it
feels a little creepy claiming to be something you're not 100% coverint
the space of, so I just moved it back to object, and aped a few handy
things, which I'll call out in the docstring. I think that's better.